basic login working
