+++ src/api/cookies.ts
@@ -1,3 +1,4 @@
+// @ts-ignore
 import Cookies from 'js-cookie';
 
 // adding cookies
