+++ src/components/molecules/LoginForm/LoginForm.pcss
@@ -1,0 +1,26 @@
+.login-form {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  justify-content: center;
+}
+
+.login-form__button {
+  margin: 0 auto;
+}
+
+.login-form__input {
+  border-radius: 4px;
+  border: solid 1px lightgrey;
+  padding: 1rem;
+  width: 100%;
+}
+
+.login-form__label {
+  padding: 0 18rem;
+}
+
+.login-form__loader {
+  display: flex;
+  justify-content: center;
+}
