+++ src/components/molecules/Navigation/Navigation.pcss
@@ -45,27 +45,3 @@
     flex-direction: row;
   }
 }
-
-.navigation__login {
-  font-family: var(--font-family-title);
-  font-size: 1rem;
-  font-weight: var(--font-weight-semibold);
-  text-decoration: none;
-  text-transform: uppercase;
-
-  &:hover,
-  &:focus,
-  &:focus-within {
-    cursor: pointer;
-    text-decoration: underline;
-  }
-}
-
-.navigation__login__button {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  gap: 5px;
-  cursor: pointer;
-}
