+++ src/components/organisms/Login/index.tsx
@@ -1,86 +1,52 @@
 'use client';
 
-import Typography from '@/components/atoms/Typography';
 import styles from './Login.css';
-import { useState } from 'react';
-import Button from '@/components/atoms/Button';
-import Loader from '@/components/atoms/Loader';
-import { addCookie } from '@/api/cookies';
+import { useEffect, useState } from 'react';
+import { hasCookie, removeCookie } from '@/api/cookies';
+import Modal from '@/components/molecules/Modal';
+import Icon from '@/components/atoms/Icon';
+import { IconLock } from '@/assets/icons';
+import LoginForm from '@/components/molecules/LoginForm';
 
 export const links = () => {
   return [{ rel: 'stylesheet', href: styles }];
 };
 
 const Login = () => {
-  const [isLoading, setLoading] = useState(false);
-
-  const tryLogin = async (event: any) => {
-    event.preventDefault();
-    setLoading(true);
-
-    const data = {
-      username: event.target.elements.username.value,
-      password: event.target.elements.password.value,
-    };
-    const JSONdata = JSON.stringify(data);
-    const options = {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSONdata,
-    };
-    const response = await fetch('/api/login', options);
-    const result = await response.json();
-    setLoading(false);
-    if (response.status !== 200) {
-      console.log('Login unsuccesfull' + response.status + '\n Error message: ' + result.data);
-      return;
-    }
-    addCookie('leader', result.token);
-    location.reload();
-  };
+  const [isLoggedIn, setIsLoggedIn] = useState(false);
+  useEffect(() => {
+    setIsLoggedIn(hasCookie('leader'));
+  }, [isLoggedIn]);
 
   return (
-    <div>
-      <form onSubmit={tryLogin} className="login__form">
-        <h2 className="t-headline-2 t-align-center">Login</h2>
-        <label htmlFor="username" className="login__label">
-          <Typography>Gebruikersnaam</Typography>
-          <input
-            className="login__input"
-            id="username"
-            name="username"
-            type="text"
-            placeholder="Gebruikersnaam"
-            autoComplete="username"
-            required
-          />
-        </label>
-        <label htmlFor="password" className="login__label">
-          <Typography>Wachtwoord</Typography>
-          <input
-            className="login__input"
-            id="password"
-            name="password"
-            type="password"
-            autoComplete="current-password"
-            placeholder="****************"
-            required
-          />
-        </label>
-        {!isLoading && (
-          <Button type="submit" className="login__button">
-            Log In
-          </Button>
-        )}
-        {isLoading && (
-          <div className="login__loader">
-            <Loader />
-          </div>
-        )}
-      </form>
-    </div>
+    <li className="login" key="login">
+      {!isLoggedIn && (
+        <Modal
+          button={
+            <div className="login__button">
+              Log in
+              <Icon icon={IconLock} size="md" title="loginLock" className="login__lock" />
+            </div>
+          }
+          modalData={<LoginForm />}
+          cardClass="login__card"
+          key="notLoggedIn"
+        />
+      )}
+      {isLoggedIn && (
+        <div
+          className="login__button"
+          onClick={() => {
+            removeCookie('leader');
+            location.reload();
+          }}
+          key="loggedIn"
+        >
+          Log uit
+          <Icon icon={IconLock} size="md" title="loginLock" className="login__lock" />
+        </div>
+      )}
+    </li>
   );
 };
